same concern with better name
